requirements.txt
aioredis==1.3.1
alabaster==0.7.12
amqp==2.6.1
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.3.1
astroid==2.5.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.0
backcall==0.2.0
billiard==3.6.3.0
black==20.8b1
boto3==1.17.39
botocore==1.20.39
celery==4.4.6
certifi==2020.12.5
cffi==1.14.5
cfgv==3.2.0
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.7
daphne==3.0.1
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.1
Django==3.1.7
django-allauth==0.44.0
django-appconf==1.0.4
django-celery-beat==2.2.0
django-compressor==2.4
django-cors-headers==3.7.0
django-coverage-plugin==1.8.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.2
django-elasticsearch-dsl==7.1.4
django-elasticsearch-dsl-drf==0.22
django-environ==0.4.5
django-extensions==3.1.1
django-model-utils==4.1.1
django-nine==0.2.4
django-redis==4.12.1
django-storages==1.11.1
django-stubs==1.7.0
django-templated-mail==1.1.1
django-timezone-field==4.1.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
docutils==0.16
drf-writable-nested==0.6.2
elasticsearch==7.12.0
elasticsearch-dsl==7.3.0
factory-boy==3.2.0
Faker==6.6.3
filelock==3.0.12
flake8==3.9.0
flake8-isort==4.0.0
future==0.18.2
graphviz==0.16
h11==0.12.0
hiredis==1.1.0
httptools==0.1.1
hyperlink==21.0.0
identify==2.2.1
idna==2.10
imagesize==1.2.0
incremental==21.3.0
iniconfig==1.1.1
install==1.3.4
ipdb==0.13.7
ipython==7.22.0
ipython-genutils==0.2.0
isort==5.8.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
kombu==4.6.11
lazy-object-proxy==1.6.0
livereload==2.6.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.2
mypy==0.812
mypy-extensions==0.4.3
nodeenv==1.5.0
oauthlib==3.1.0
packaging==20.9
parso==0.8.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
pluggy==0.13.1
pre-commit==2.11.1
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydot==1.4.2
pydotplus==2.0.2
pyflakes==2.3.1
Pygments==2.8.1
PyJWT==2.0.1
pylint==2.7.2
pylint-celery==0.3
pylint-django==2.4.2
pylint-plugin-utils==0.6
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.1.0
pytest-sugar==0.9.4
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.16.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
rcssmin==1.0.6
redis==3.5.3
regex==2021.3.17
requests==2.25.1
requests-oauthlib==1.3.0
rjsmin==1.1.0
s3transfer==0.3.6
service-identity==18.1.0
six==1.15.0
snowballstemmer==2.1.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
Sphinx==3.5.3
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
termcolor==1.1.0
testfixtures==6.17.1
text-unidecode==1.3
toml==0.10.2
tornado==6.1
traitlets==5.0.5
Twisted==21.2.0
txaio==21.2.1
typed-ast==1.4.2
typing==3.7.4.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
uuid==1.30
uvicorn==0.13.4
uvloop==0.15.2
vine==1.3.0
virtualenv==20.4.3
watchgod==0.6
wcwidth==0.2.5
websockets==8.1
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
zope.interface==5.3.0
